requirements.txt
pytube
openai-whisper
transformers
torch
yt-dlp
python-dotenv
ffmpeg-python
tf-keras
sentencepiece